extract transcript code
